requirements.txt
face_recognition==1.3.0
uvicorn==0.19.0
numpy==1.21.6
fastapi==0.88.0
pydantic==1.10.2
pillow==9.3.0
requests==2.27.1
opencv-python==4.7.0.68
python-multipart==0.0.5